Move optimizes director out from tensor to make it universal.
