+++ mars/executor.py
@@ -432,7 +432,7 @@ class Executor(object):
 
     def _preprocess(self, graph, keys):
         # TODO(xuye.qin): make an universal optimzier
-        from .tensor.optimizes.core import Optimizer
+        from .optimizes.core import Optimizer
 
         Optimizer(graph, self._engine).optimize(keys=keys)
         return graph
